upload-worker/requirements.txt
blinker==1.8.2
boto3==1.34.150
botocore==1.34.150
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.1
playwright==1.45.1
pluggy==1.5.0
pyee==11.1.0
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
